+++ src/core/core-config.c
@@ -3643,10 +3643,12 @@ config_weechat_init_options (void)
             /* TRANSLATORS: string "${color:xxx}" must NOT be translated. */
             N_("message displayed when the day has changed, with two dates "
                "displayed (between two messages); the second date specifiers "
-               "must start with two \"%\" because strftime is called two "
-               "times on this string (see man strftime for date/time "
-               "specifiers) (note: content is evaluated, so you can use "
-               "colors with format \"${color:xxx}\", see /help eval)"),
+               "must start with two \"%\" because the date/time specifiers "
+               "are replaced two times in this string (see man strftime for "
+               "date/time specifiers, extra specifiers are supported, see "
+               "function util_strftimeval in Plugin API reference) (note: "
+               "content is evaluated, so you can use colors with format "
+               "\"${color:xxx}\", see /help eval)"),
             NULL, 0, 0, "-- %%a, %%d %%b %%Y (%a, %d %b %Y) --", NULL, 0,
             NULL, NULL, NULL,
             &config_change_buffers, NULL, NULL,